# Accelerate gamma_adjust for uint8 images