--- a/skimage/exposure/exposure.py
+++ b/skimage/exposure/exposure.py
@@ -436,6 +436,16 @@ def _assert_non_negative(image):
                          'skimage.exposure.rescale_intensity.')
 
 
+def _adjust_gamma_u8(image, gamma, gain):
+    """LUT based implmentation of gamma adjustement.
+
+    """
+    lut_size = image.max() + 1
+    # lut = (255 * gain * (np.linspace(0, 1, lut_size) ** gamma))
+    lut = (255 * gain * ((np.arange(0, lut_size) / 255) ** gamma))
+    return lut.astype('uint8')[image]
+
+
 def adjust_gamma(image, gamma=1, gain=1):
     """Performs Gamma Correction on the input image.
 
@@ -482,11 +492,15 @@ def adjust_gamma(image, gamma=1, gain=1):
     >>> image.mean() > gamma_corrected.mean()
     True
     """
-    _assert_non_negative(image)
-    dtype = image.dtype.type
-
     if gamma < 0:
         raise ValueError("Gamma should be a non-negative real number.")
+
+    dtype = image.dtype.type
+
+    if dtype is np.uint8:
+        return _adjust_gamma_u8(image, gamma, gain)
+
+    _assert_non_negative(image)
 
     scale = float(dtype_limits(image, True)[1] - dtype_limits(image, True)[0])