@sciunto 2020-09-10
Nice! I think it would be too much work to add a benchmark for this function: https://github.com/scikit-image/scikit-image/blob/master/benchmarks/benchmark_exposure.py