@emmanuelle 2020-09-14
> There seems to be something wrong with Travis, but it's not specific to this PR, it also happens on recent commits in other PRs.

There has been a release of dask on Friday, maybe this is why? It's probably related to one of the optional deps since only those builds are failing.